Refactorizacion de la pagina de blogs Se asilaron los datos de prueba y se creo un componente para contener el contenido de los blogs
